data-mgmt/requirements.txt
SQLAlchemy==1.3.12
SQLAlchemy-Utils==0.36.1
alembic==1.3.2
requests==2.22.0
bs4==0.0.1
psycopg2-binary==2.8.4
python-slugify==4.0.0
